added create user client and server

+++ src/main/java/com/example/wbdvsu119serverjava/models/User.java
@@ -1,15 +1,15 @@
 package com.example.wbdvsu119serverjava.models;
 
 public class User {
-    private int id;
+    private long id;
     private String username;
     private String password;
     private String firstName;
     private String lastName;
-    public int getId() {
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
     public String getUsername() {
@@ -46,7 +46,7 @@ public class User {
     public User() {
         super();
     }
-    public User(int id, String username, String password, String firstName, String lastName) {
+    public User(long id, String username, String password, String firstName, String lastName) {
         super();
         this.id = id;
         this.username = username;
